requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
gTTS==2.2.2
idna==2.10
numpy==1.20.2
opencv-python==4.5.1.48
picamera==1.13
Pillow==8.2.0
python-vlc==3.0.12118
requests==2.25.1
six==1.16.0
urllib3==1.26.4
